pyproject.toml
[build-system]
requires = ['setuptools>=61.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'zx'
version = '0.11.0'
description = 'ZX Spectrum emulation framework'
readme = 'README.md'
license = {text = 'MIT'}
authors = [
  {name = 'Ivan Kosarev', email = 'mail@ivankosarev.com'}
]
dependencies = [
  'pysdl2',
  'pysdl2-dll',
  'numpy',
  'evdev; sys_platform == "linux"'
]
requires-python = '>=3.7'

[project.urls]
Homepage = 'https://github.com/kosarev/zx/'

[project.scripts]
zx = 'zx:main'

[tool.setuptools]
packages = ['zx']

[tool.setuptools.package-data]
zx = ['roms/*', 'py.typed']
